Validated DDD field, corrected imports, renamed pages to match the style guide.
