+++ src/app/app.component.spec.ts
@@ -1,9 +1,11 @@
 
 import { async, TestBed } from '@angular/core/testing';
 import { IonicModule, Platform } from 'ionic-angular';
-import { StatusBar } from '@ionic-native/status-bar';
 import { SplashScreen } from '@ionic-native/splash-screen';
+import { StatusBar } from '@ionic-native/status-bar';
+
 import { MyApp } from './app.component';
+
 import {
   PlatformMock,
   StatusBarMock,
