+++ src/app/app.component.ts
@@ -1,17 +1,9 @@
 import { Component, ViewChild } from '@angular/core';
-import { Nav, Platform, MenuController } from 'ionic-angular';
-import { StatusBar } from '@ionic-native/status-bar';
+import { MenuController, Nav, Platform } from 'ionic-angular';
 import { SplashScreen } from '@ionic-native/splash-screen';
+import { StatusBar } from '@ionic-native/status-bar';
 
-/*import { AboutPage } from '../pages/about/about';
-import { ActivitiesListPage } from '../pages/activities-list/activities-list';
-import { LoginPage } from '../pages/login/login';
-import { ProfilePage} from '../pages/profile/profile';
-import { RegisterPage } from '../pages/register/register';
-import { SettingsPage } from '../pages/settings/settings';*/
 import { WelcomePage } from '../pages/welcome/welcome';
-//import { TabsPage } from '../pages/tabs/tabs';
-
 
 @Component({
   templateUrl: 'app.html'
@@ -21,24 +13,10 @@ export class MyApp {
   @ViewChild(Nav) nav: Nav;
 
   rootPage: any = WelcomePage;
-  //pages: Array<{title: string, component: any}>;
 
   constructor(public platform: Platform, public statusBar: StatusBar, public splashScreen: SplashScreen, public menu: MenuController) {
     this.initializeApp();
-
-    /* Menu antigo
-    this.pages = [
-      { title: 'Bem-Vindo', component: WelcomePage },
-      { title: 'Login', component: LoginPage },
-      { title: 'Registro', component: RegisterPage },
-      { title: 'Lista de Atividades', component: ActivitiesListPage },
-      { title: 'Atividades', component: ActivitiesListPage },
-      { title: 'Perfil', component: ProfilePage },
-      { title: 'Sobre', component: AboutPage },
-      { title: 'Configurações', component: SettingsPage }
-    ];*/
   }
-
 
   initializeApp() {
     this.platform.ready().then(() => {
