+++ src/app/app.module.ts
@@ -1,52 +1,50 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { ErrorHandler, NgModule } from '@angular/core';
+import { HttpClientModule } from '@angular/common/http'
 import { IonicApp, IonicErrorHandler, IonicModule } from 'ionic-angular';
 import { IonicStorageModule } from '@ionic/storage';
+import { SplashScreen } from '@ionic-native/splash-screen';
+import { StatusBar } from '@ionic-native/status-bar';
+
+import { MyApp } from './app.component';
+import { RestActivityProvider } from '../providers/rest-activity';
+import { RestUserProvider } from '../providers/rest-user';
+import { StorageService } from '../providers/storage.service';
 
 import { AboutPage } from '../pages/about/about';
-import { ActivityDetailsPage } from '../pages/activity-details/activity-details';
 import { ActivitiesListPage } from '../pages/activities-list/activities-list';
-import { MyApp } from './app.component';
+import { ActivityDetailsPage } from '../pages/activity-details/activity-details';
+import { HelpPage } from '../pages/help/help';
+import { ListActivityPage } from '../pages/list-activity/list-activity';
+import { ListUserPage } from '../pages/listuser/listuser';
 import { LoginPage } from '../pages/login/login';
+import { MenuAppPage } from '../pages/menu-app/menu-app';
+import { NewsPage } from '../pages/news/news';
 import { ProfilePage } from '../pages/profile/profile';
 import { RegisterPage } from '../pages/register/register';
 import { SettingsPage } from '../pages/settings/settings';
-import { WelcomePage } from '../pages/welcome/welcome';
-import { ListUserPage } from '../pages/listuser/listuser';
 import { TabsPage } from '../pages/tabs/tabs';
-import { NewsPage } from '../pages/news/news';
-import { MenuAppPage } from '../pages/menuApp/menuApp';
-import { ListActivityPage } from '../pages/list-activity/list-activity';
-import { HelpPage } from '../pages/help/help';
+import { WelcomePage } from '../pages/welcome/welcome';
 
-import { StatusBar } from '@ionic-native/status-bar';
-import { SplashScreen } from '@ionic-native/splash-screen';
-import { RestActivityProvider } from '../providers/rest-activity';
-
-import { HttpClientModule } from '@angular/common/http'
-import { RestUserProvider } from '../providers/rest-user';
-import { StorageService } from '../providers/storage.service';
 
 
 @NgModule({
   declarations: [
-    MyApp,
-    WelcomePage,
-    LoginPage,
-    RegisterPage,
-    ListUserPage,
+    AboutPage,
     ActivitiesListPage,
     ActivityDetailsPage,
-    ProfilePage,
-    AboutPage,
-    SettingsPage,
-    ListUserPage,
-    TabsPage,
-    NewsPage,
-    MenuAppPage,
+    HelpPage,
     ListActivityPage,
-    HelpPage
-    
+    ListUserPage,
+    LoginPage,
+    MenuAppPage,
+    MyApp,
+    NewsPage,
+    ProfilePage,
+    RegisterPage,
+    SettingsPage,
+    TabsPage,
+    WelcomePage
   ],
 
   imports: [
@@ -59,29 +57,28 @@ import { StorageService } from '../providers/storage.service';
   bootstrap: [IonicApp],
 
   entryComponents: [
-    MyApp,
-    WelcomePage,
-    LoginPage,
-    RegisterPage,
-    ListUserPage,
+    AboutPage,
     ActivitiesListPage,
     ActivityDetailsPage,
-    ProfilePage,
-    AboutPage,
-    SettingsPage,
-    ListUserPage,
-    TabsPage,
-    NewsPage,
-    MenuAppPage,
+    HelpPage,
     ListActivityPage,
-    HelpPage
+    ListUserPage,
+    LoginPage,
+    MenuAppPage,
+    MyApp,
+    NewsPage,
+    ProfilePage,
+    RegisterPage,
+    SettingsPage,
+    TabsPage,
+    WelcomePage
   ],
 
   providers: [
-    StatusBar,
     RestActivityProvider,
-    SplashScreen,
     RestUserProvider,
+    SplashScreen,
+    StatusBar,
     StorageService,
     {provide: ErrorHandler, useClass: IonicErrorHandler},
   ]
