+++ src/pages/activity-details/activity-details.ts
@@ -2,6 +2,7 @@ import { Component } from '@angular/core';
 import { NavController, NavParams } from 'ionic-angular';
 
 import { RestActivityProvider } from '../../providers/rest-activity';
+
 import { ActivitiesListPage } from '../activities-list/activities-list';
 
 @Component({
