+++ src/pages/menu-app/menu-app.scss
@@ -1,4 +1,4 @@
-page-menuApp {
+page-menu-app {
 
   $lightblue-color: #0565c0;
   $blue-color: #084887;
