+++ src/pages/welcome/welcome.ts
@@ -1,5 +1,6 @@
 import { Component } from '@angular/core';
 import { NavController, NavParams } from 'ionic-angular';
+
 import { LoginPage } from '../login/login';
 
 @Component({
